refactoring: change import to UI dir
